// frontend/src/api/models/TaskResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TaskResponse
 */
export interface TaskResponse {
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    description?: string | null;
    /**
     * 
     * @type {number}
     * @memberof TaskResponse
     */
    courseId: number;
    /**
     * 
     * @type {boolean}
     * @memberof TaskResponse
     */
    isActive?: boolean;
    /**
     * 
     * @type {number}
     * @memberof TaskResponse
     */
    taskId: number;
}

/**
 * Check if a given object implements the TaskResponse interface.
 */
export function instanceOfTaskResponse(value: object): value is TaskResponse {
    if (!('courseId' in value) || value['courseId'] === undefined) return false;
    if (!('taskId' in value) || value['taskId'] === undefined) return false;
    return true;
}

export function TaskResponseFromJSON(json: any): TaskResponse {
    return TaskResponseFromJSONTyped(json, false);
}

export function TaskResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TaskResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'courseId': json['course_id'],
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
        'taskId': json['task_id'],
    };
}

export function TaskResponseToJSON(json: any): TaskResponse {
    return TaskResponseToJSONTyped(json, false);
}

export function TaskResponseToJSONTyped(value?: TaskResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'description': value['description'],
        'course_id': value['courseId'],
        'is_active': value['isActive'],
        'task_id': value['taskId'],
    };
}